atualizado exercícios

+++ desafios/desafio007.py
@@ -1,4 +1,6 @@
-n1 = int(input('Primeira nota: '))
-n2 = int(input('Segunda nota: '))
+n1 = float(input('Primeira nota: '))
+n2 = float(input('Segunda nota: '))
 media = (n1 + n2) / 2
 print(f'A média das notas é {media}')
+
+# print(f'A média das notas é {(n1 + n2) / 2}')
